requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-cloud-texttospeech==2.15.0
google-auth==2.27.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0
python-dotenv==1.0.0
